refactor(MobilityProfileSelector): Extract mobility profile dropdown
